requirements.txt
python-dotenv
googlesearch-python
streamlit
google-generativeai
